Revert back to 1.20.1
Undo changes to bring the mod back to the most recent version compatible with 1.20.1.

+++ src/main/java/sirttas/elementalcraft/block/source/SourceSourceTraitHolder.java
@@ -15,8 +15,8 @@ public class SourceSourceTraitHolder extends SourceTraitHolder {
     }
 
     @Override
-    public float getRecoverRate() {
-        return super.getRecoverRate() + (source.isStabilized() ? 20F : 0F);
+    public int getRecoverRate() {
+        return super.getRecoverRate() + (source.isStabilized() ? 20 : 0);
     }
 
     @Override
